fix: split name resolver setup into its own service
The creation of /etc/resolv.conf should wait for
network-online.target, but setup_host_network is supposed to be run
before network is setup. Thus we split the setup of name resolvers
into its own service to fix this issue.

Ref bsc#1246513
