+++ setup.py
@@ -25,6 +25,7 @@ config = {
             'suse-migration-mount-system=suse_migration_services.units.mount_system:main',
             'suse-migration-ssh-keys=suse_migration_services.units.ssh_keys:main',
             'suse-migration-setup-host-network=suse_migration_services.units.setup_host_network:main',
+            'suse-migration-setup-name-resolver=suse_migration_services.units.setup_name_resolver:main',
             'suse-migration-btrfs-snapshot-pre-migration=suse_migration_services.units.btrfs_snapshot_pre_migration:main',
             'suse-migration-btrfs-snapshot-post-migration=suse_migration_services.units.btrfs_snapshot_post_migration:main',
             'suse-migration-prepare=suse_migration_services.units.prepare:main',
